requirements.txt
PySide6==6.6.1
onnxruntime==1.17.1
Pillow==10.1.0
numpy==1.24.3
PyInstaller==6.3.0

